update seek func

+++ Source/JukeboxItem.swift
@@ -48,12 +48,12 @@ open class JukeboxItem: NSObject {
     /**
      The starting time on playing.
      */
-    public var startTime: Int = 0
+    public var startTime: Double = 0
     
     /**
      If this is non-nil, item plays at this time. Takes precedence over `startTime`.
     */
-    public var continueTime: Int?
+    public var continueTime: Double?
     
     /**
      The adjusted volume for this item. Between 0.0 to 1.0
